CocosBuilder migrated to cocos2d v2.0-beta2

+++ CocosBuilder/libs/cocos2d/CCParticleExamples.m
@@ -2,17 +2,18 @@
  * cocos2d for iPhone: http://www.cocos2d-iphone.org
  *
  * Copyright (c) 2008-2010 Ricardo Quesada
- * 
+ * Copyright (c) 2011 Zynga Inc.
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -39,7 +40,7 @@
 	return [self initWithTotalParticles:1500];
 }
 
--(id) initWithTotalParticles:(int)p
+-(id) initWithTotalParticles:(NSUInteger)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 		// duration
@@ -50,7 +51,7 @@
 
 		// Gravity Mode: gravity
 		self.gravity = ccp(0,-90);
-		
+
 		// Gravity Mode:  radial
 		self.radialAccel = 0;
 		self.radialAccelVar = 0;
@@ -58,22 +59,22 @@
 		//  Gravity Mode: speed of particles
 		self.speed = 180;
 		self.speedVar = 50;
-		
+
 		// emitter position
 		CGSize winSize = [[CCDirector sharedDirector] winSize];
 		self.position = ccp(winSize.width/2, winSize.height/2);
-		
+
 		// angle
 		angle = 90;
 		angleVar = 20;
-				
+
 		// life of particles
 		life = 3.5f;
 		lifeVar = 1;
-			
+
 		// emits per frame
 		emissionRate = totalParticles/life;
-		
+
 		// color of particles
 		startColor.r = 0.5f;
 		startColor.g = 0.5f;
@@ -91,7 +92,7 @@
 		endColorVar.g = 0.1f;
 		endColorVar.b = 0.1f;
 		endColorVar.a = 0.2f;
-		
+
 		// size, in pixels
 		startSize = 8.0f;
 		startSizeVar = 2.0f;
@@ -102,7 +103,7 @@
 		// additive
 		self.blendAdditive = NO;
 	}
-	
+
 	return self;
 }
 @end
@@ -116,7 +117,7 @@
 	return [self initWithTotalParticles:250];
 }
 
--(id) initWithTotalParticles:(int) p
+-(id) initWithTotalParticles:(NSUInteger) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 
@@ -128,29 +129,29 @@
 
 		// Gravity Mode: gravity
 		self.gravity = ccp(0,0);
-		
+
 		// Gravity Mode: radial acceleration
 		self.radialAccel = 0;
 		self.radialAccelVar = 0;
-		
+
 		// Gravity Mode: speed of particles
 		self.speed = 60;
-		self.speedVar = 20;		
-		
+		self.speedVar = 20;
+
 		// starting angle
 		angle = 90;
 		angleVar = 10;
-		
+
 		// emitter position
 		CGSize winSize = [[CCDirector sharedDirector] winSize];
 		self.position = ccp(winSize.width/2, 60);
 		posVar = ccp(40, 20);
-		
+
 		// life of particles
 		life = 3;
 		lifeVar = 0.25f;
-		
-			
+
+
 		// size, in pixels
 		startSize = 54.0f;
 		startSizeVar = 10.0f;
@@ -158,7 +159,7 @@
 
 		// emits per frame
 		emissionRate = totalParticles/life;
-		
+
 		// color of particles
 		startColor.r = 0.76f;
 		startColor.g = 0.25f;
@@ -176,13 +177,13 @@
 		endColorVar.g = 0.0f;
 		endColorVar.b = 0.0f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
-		
+
 		// additive
 		self.blendAdditive = YES;
 	}
-		
+
 	return self;
 }
 @end
@@ -196,44 +197,44 @@
 	return [self initWithTotalParticles:350];
 }
 
--(id) initWithTotalParticles:(int) p
+-(id) initWithTotalParticles:(NSUInteger) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 
 		// additive
 		self.blendAdditive = YES;
-			
+
 		// duration
 		duration = kCCParticleDurationInfinity;
-		
+
 		// Gravity Mode
 		self.emitterMode = kCCParticleModeGravity;
-		
+
 		// Gravity Mode: gravity
 		self.gravity = ccp(0,0);
-		
+
 		// Gravity mode: radial acceleration
 		self.radialAccel = 0;
 		self.radialAccelVar = 0;
-		
+
 		// Gravity mode: speed of particles
 		self.speed = 20;
 		self.speedVar = 5;
-				
-		
+
+
 		// angle
 		angle = 90;
 		angleVar = 360;
-		
+
 		// emitter position
 		CGSize winSize = [[CCDirector sharedDirector] winSize];
 		self.position = ccp(winSize.width/2, winSize.height/2);
 		posVar = CGPointZero;
-		
+
 		// life of particles
 		life = 1;
 		lifeVar = 0.5f;
-		
+
 		// size, in pixels
 		startSize = 30.0f;
 		startSizeVar = 10.0f;
@@ -241,7 +242,7 @@
 
 		// emits per seconds
 		emissionRate = totalParticles/life;
-		
+
 		// color of particles
 		startColor.r = 0.76f;
 		startColor.g = 0.25f;
@@ -259,10 +260,10 @@
 		endColorVar.g = 0.0f;
 		endColorVar.b = 0.0f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
 	}
-		
+
 	return self;
 }
 @end
@@ -276,7 +277,7 @@
 	return [self initWithTotalParticles:200];
 }
 
--(id) initWithTotalParticles:(int)p
+-(id) initWithTotalParticles:(NSUInteger)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 
@@ -288,40 +289,40 @@
 
 		// Gravity Mode: gravity
 		self.gravity = ccp(0,0);
-		
+
 		// Gravity Mode: speed of particles
 		self.speed = 60;
 		self.speedVar = 10;
-			
+
 		// Gravity Mode: radial
 		self.radialAccel = -80;
 		self.radialAccelVar = 0;
-		
+
 		// Gravity Mode: tagential
 		self.tangentialAccel = 80;
 		self.tangentialAccelVar = 0;
-		
+
 		// angle
 		angle = 90;
 		angleVar = 360;
-		
+
 		// emitter position
 		CGSize winSize = [[CCDirector sharedDirector] winSize];
 		self.position = ccp(winSize.width/2, winSize.height/2);
 		posVar = CGPointZero;
-		
+
 		// life of particles
 		life = 4;
 		lifeVar = 1;
-		
+
 		// size, in pixels
 		startSize = 37.0f;
 		startSizeVar = 10.0f;
 		endSize = kCCParticleStartSizeEqualToEndSize;
-		
+
 		// emits per second
 		emissionRate = totalParticles/life;
-		
+
 		// color of particles
 		startColor.r = 0.12f;
 		startColor.g = 0.25f;
@@ -339,13 +340,13 @@
 		endColorVar.g = 0.0f;
 		endColorVar.b = 0.0f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
 
 		// additive
 		self.blendAdditive = YES;
 	}
-	
+
 	return self;
 }
 @end
@@ -359,10 +360,10 @@
 	return [self initWithTotalParticles:250];
 }
 
--(id) initWithTotalParticles:(int) p
+-(id) initWithTotalParticles:(NSUInteger) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
-	
+
 		// duration
 		duration = kCCParticleDurationInfinity;
 
@@ -371,15 +372,15 @@
 
 		// Gravity Mode: gravity
 		self.gravity = ccp(0,0);
-		
+
 		// Gravity Mode: speed of particles
 		self.speed = 80;
 		self.speedVar = 10;
-		
+
 		// Gravity Mode: radial
 		self.radialAccel = -60;
 		self.radialAccelVar = 0;
-		
+
 		// Gravity Mode: tagential
 		self.tangentialAccel = 15;
 		self.tangentialAccelVar = 0;
@@ -387,16 +388,16 @@
 		// angle
 		angle = 90;
 		angleVar = 360;
-		
+
 		// emitter position
 		CGSize winSize = [[CCDirector sharedDirector] winSize];
 		self.position = ccp(winSize.width/2, winSize.height/2);
 		posVar = CGPointZero;
-		
+
 		// life of particles
 		life = 4;
 		lifeVar = 1;
-		
+
 		// size, in pixels
 		startSize = 30.0f;
 		startSizeVar = 10.0f;
@@ -404,7 +405,7 @@
 
 		// emits per second
 		emissionRate = totalParticles/life;
-		
+
 		// color of particles
 		startColor.r = 0.50f;
 		startColor.g = 0.50f;
@@ -422,13 +423,13 @@
 		endColorVar.g = 0.0f;
 		endColorVar.b = 0.0f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
 
 		// additive
 		self.blendAdditive = YES;
 	}
-		
+
 	return self;
 }
 @end
@@ -442,13 +443,13 @@
 	return [self initWithTotalParticles:150];
 }
 
--(id) initWithTotalParticles:(int) p
+-(id) initWithTotalParticles:(NSUInteger) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
 
 		// duration
 		duration = kCCParticleDurationInfinity;
-		
+
 		// Gravity Mode
 		self.emitterMode = kCCParticleModeGravity;
 
@@ -458,28 +459,28 @@
 		// Gravity Mode: speed of particles
 		self.speed = 15;
 		self.speedVar = 5;
-		
+
 		// Gravity Mode: radial
 		self.radialAccel = 0;
 		self.radialAccelVar = 0;
-		
+
 		// Gravity Mode: tagential
 		self.tangentialAccel = 0;
 		self.tangentialAccelVar = 0;
-		
+
 		// angle
 		angle = 90;
 		angleVar = 360;
-		
+
 		// emitter position
 		CGSize winSize = [[CCDirector sharedDirector] winSize];
 		self.position = ccp(winSize.width/2, winSize.height/2);
 		posVar = CGPointZero;
-		
+
 		// life of particles
 		life = 2;
 		lifeVar = 1;
-		
+
 		// size, in pixels
 		startSize = 60.0f;
 		startSizeVar = 10.0f;
@@ -487,7 +488,7 @@
 
 		// emits per second
 		emissionRate = totalParticles/life;
-		
+
 		// color of particles
 		startColor.r = 0.2f;
 		startColor.g = 0.4f;
@@ -505,13 +506,13 @@
 		endColorVar.g = 0.0f;
 		endColorVar.b = 0.0f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
-		
+
 		// additive
 		self.blendAdditive = YES;
 	}
-	
+
 	return self;
 }
 @end
@@ -525,44 +526,44 @@
 	return [self initWithTotalParticles:500];
 }
 
--(id) initWithTotalParticles:(int) p
+-(id) initWithTotalParticles:(NSUInteger) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
-	
+
 		// duration
 		duration = kCCParticleDurationInfinity;
 
 		// Gravity Mode
 		self.emitterMode = kCCParticleModeGravity;
-		
+
 		// Gravity Mode: gravity
 		self.gravity = ccp(0,0);
-		
+
 		// Gravity Mode: speed of particles
 		self.speed = 150;
 		self.speedVar = 0;
-		
+
 		// Gravity Mode: radial
 		self.radialAccel = -380;
 		self.radialAccelVar = 0;
-		
+
 		// Gravity Mode: tagential
 		self.tangentialAccel = 45;
 		self.tangentialAccelVar = 0;
-		
+
 		// angle
 		angle = 90;
 		angleVar = 0;
-		
+
 		// emitter position
 		CGSize winSize = [[CCDirector sharedDirector] winSize];
 		self.position = ccp(winSize.width/2, winSize.height/2);
 		posVar = CGPointZero;
-		
+
 		// life of particles
 		life = 12;
 		lifeVar = 0;
-		
+
 		// size, in pixels
 		startSize = 20.0f;
 		startSizeVar = 0.0f;
@@ -570,7 +571,7 @@
 
 		// emits per second
 		emissionRate = totalParticles/life;
-		
+
 		// color of particles
 		startColor.r = 0.5f;
 		startColor.g = 0.5f;
@@ -588,13 +589,13 @@
 		endColorVar.g = 0.5f;
 		endColorVar.b = 0.5f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
 
 		// additive
 		self.blendAdditive = NO;
 	}
-	
+
 	return self;
 }
 @end
@@ -608,43 +609,43 @@
 	return [self initWithTotalParticles:700];
 }
 
--(id) initWithTotalParticles:(int)p
+-(id) initWithTotalParticles:(NSUInteger)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
-	
+
 		// duration
 		duration = 0.1f;
-		
+
 		self.emitterMode = kCCParticleModeGravity;
 
 		// Gravity Mode: gravity
 		self.gravity = ccp(0,0);
-		
+
 		// Gravity Mode: speed of particles
 		self.speed = 70;
 		self.speedVar = 40;
-		
+
 		// Gravity Mode: radial
 		self.radialAccel = 0;
 		self.radialAccelVar = 0;
-		
+
 		// Gravity Mode: tagential
 		self.tangentialAccel = 0;
 		self.tangentialAccelVar = 0;
-		
+
 		// angle
 		angle = 90;
 		angleVar = 360;
-				
+
 		// emitter position
 		CGSize winSize = [[CCDirector sharedDirector] winSize];
 		self.position = ccp(winSize.width/2, winSize.height/2);
 		posVar = CGPointZero;
-		
+
 		// life of particles
 		life = 5.0f;
 		lifeVar = 2;
-		
+
 		// size, in pixels
 		startSize = 15.0f;
 		startSizeVar = 10.0f;
@@ -652,7 +653,7 @@
 
 		// emits per second
 		emissionRate = totalParticles/duration;
-		
+
 		// color of particles
 		startColor.r = 0.7f;
 		startColor.g = 0.1f;
@@ -670,13 +671,13 @@
 		endColorVar.g = 0.5f;
 		endColorVar.b = 0.5f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
 
 		// additive
 		self.blendAdditive = NO;
 	}
-	
+
 	return self;
 }
 @end
@@ -690,40 +691,40 @@
 	return [self initWithTotalParticles:200];
 }
 
--(id) initWithTotalParticles:(int) p
+-(id) initWithTotalParticles:(NSUInteger) p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
-	
+
 		// duration
 		duration = kCCParticleDurationInfinity;
-		
+
 		// Emitter mode: Gravity Mode
 		self.emitterMode = kCCParticleModeGravity;
-		
+
 		// Gravity Mode: gravity
 		self.gravity = ccp(0,0);
 
 		// Gravity Mode: radial acceleration
 		self.radialAccel = 0;
 		self.radialAccelVar = 0;
-		
+
 		// Gravity Mode: speed of particles
 		self.speed = 25;
 		self.speedVar = 10;
-		
+
 		// angle
 		angle = 90;
 		angleVar = 5;
-		
+
 		// emitter position
 		CGSize winSize = [[CCDirector sharedDirector] winSize];
 		self.position = ccp(winSize.width/2, 0);
 		posVar = ccp(20, 0);
-		
+
 		// life of particles
 		life = 4;
 		lifeVar = 1;
-		
+
 		// size, in pixels
 		startSize = 60.0f;
 		startSizeVar = 10.0f;
@@ -731,7 +732,7 @@
 
 		// emits per frame
 		emissionRate = totalParticles/life;
-		
+
 		// color of particles
 		startColor.r = 0.8f;
 		startColor.g = 0.8f;
@@ -749,13 +750,13 @@
 		endColorVar.g = 0.0f;
 		endColorVar.b = 0.0f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
-		
+
 		// additive
 		self.blendAdditive = NO;
 	}
-	
+
 	return self;
 }
 @end
@@ -766,38 +767,38 @@
 	return [self initWithTotalParticles:700];
 }
 
--(id) initWithTotalParticles:(int)p
+-(id) initWithTotalParticles:(NSUInteger)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
-	
+
 		// duration
 		duration = kCCParticleDurationInfinity;
-		
+
 		// set gravity mode.
 		self.emitterMode = kCCParticleModeGravity;
 
 		// Gravity Mode: gravity
 		self.gravity = ccp(0,-1);
-		
+
 		// Gravity Mode: speed of particles
 		self.speed = 5;
 		self.speedVar = 1;
-		
+
 		// Gravity Mode: radial
 		self.radialAccel = 0;
 		self.radialAccelVar = 1;
-		
+
 		// Gravity mode: tagential
 		self.tangentialAccel = 0;
 		self.tangentialAccelVar = 1;
-		
+
 		// emitter position
 		self.position = (CGPoint) {
 			[[CCDirector sharedDirector] winSize].width / 2,
 			[[CCDirector sharedDirector] winSize].height + 10
 		};
 		posVar = ccp( [[CCDirector sharedDirector] winSize].width / 2, 0 );
-		
+
 		// angle
 		angle = -90;
 		angleVar = 5;
@@ -805,7 +806,7 @@
 		// life of particles
 		life = 45;
 		lifeVar = 15;
-		
+
 		// size, in pixels
 		startSize = 10.0f;
 		startSizeVar = 5.0f;
@@ -813,7 +814,7 @@
 
 		// emits per second
 		emissionRate = 10;
-		
+
 		// color of particles
 		startColor.r = 1.0f;
 		startColor.g = 1.0f;
@@ -831,13 +832,13 @@
 		endColorVar.g = 0.0f;
 		endColorVar.b = 0.0f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
-		
+
 		// additive
 		self.blendAdditive = NO;
 	}
-		
+
 	return self;
 }
 @end
@@ -848,22 +849,22 @@
 	return [self initWithTotalParticles:1000];
 }
 
--(id) initWithTotalParticles:(int)p
+-(id) initWithTotalParticles:(NSUInteger)p
 {
 	if( (self=[super initWithTotalParticles:p]) ) {
-	
+
 		// duration
 		duration = kCCParticleDurationInfinity;
-		
+
 		self.emitterMode = kCCParticleModeGravity;
 
 		// Gravity Mode: gravity
 		self.gravity = ccp(10,-10);
-		
+
 		// Gravity Mode: radial
 		self.radialAccel = 0;
 		self.radialAccelVar = 1;
-		
+
 		// Gravity Mode: tagential
 		self.tangentialAccel = 0;
 		self.tangentialAccelVar = 1;
@@ -871,23 +872,23 @@
 		// Gravity Mode: speed of particles
 		self.speed = 130;
 		self.speedVar = 30;
-		
+
 		// angle
 		angle = -90;
 		angleVar = 5;
-		
-		
+
+
 		// emitter position
 		self.position = (CGPoint) {
 			[[CCDirector sharedDirector] winSize].width / 2,
 			[[CCDirector sharedDirector] winSize].height
 		};
 		posVar = ccp( [[CCDirector sharedDirector] winSize].width / 2, 0 );
-		
+
 		// life of particles
 		life = 4.5f;
 		lifeVar = 0;
-		
+
 		// size, in pixels
 		startSize = 4.0f;
 		startSizeVar = 2.0f;
@@ -895,7 +896,7 @@
 
 		// emits per second
 		emissionRate = 20;
-		
+
 		// color of particles
 		startColor.r = 0.7f;
 		startColor.g = 0.8f;
@@ -913,13 +914,13 @@
 		endColorVar.g = 0.0f;
 		endColorVar.b = 0.0f;
 		endColorVar.a = 0.0f;
-		
+
 		self.texture = [[CCTextureCache sharedTextureCache] addImage: @"fire.png"];
-		
+
 		// additive
 		self.blendAdditive = NO;
 	}
-	
+
 	return self;
 }
 @end
